Correction du bug des images n'apparaissant pas.
Ce bug courant est rencontré quand le cache de l'application est effacé et
qu'elle n'est pas redémarrée.
Précédemment, l'application ne vérifiait l'état de son cache qu'au démarrage.
Si le cache était effacé mais que l'application était toujours en cours
d'exécution, elle essayait toujours de recharger les images depuis le cache
alors effacé.

Ce problème est devenu de plus en plus fréquent car les appareils actuels
effacent régulièrement les caches des applications en fonction de l'espace
disque. Sur les appareils avec moins d'1 Go d'espace disponible, Android
peut effacer le cache quasi-systématiquement, causant le problème en
permanence.

Les modifications dans cette révision lancent un rafraîchissement de l'état
du cache de l'application quand un topic est (re)chargé ou qu'un lien
permanent vers un message est ouvert. Cette solution devrait gérer
convenablement les cas où le cache est effacé par le système pendant que
l'application fonctionne.
